Make session and domain directories on start
On the first start of a session, create the session directory and the
per-domain subfolders. This fixes incoherent behaviour that the session
folder was only created for local sessions with kernel tracing enabled.
No folder was created if a session was streamed to a relay and produced
nothing, or if the session only had UST events enabled and no
application started.

Now the first time a session is started, the session and per-domain
directories will always exist regardless of the activity and if it is a
local or remote session.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>